Navbar Component + Theme - Separated navbar into its own reusable component
- Moved theme to pages.js so it gets passed down
